refactor: move to components directory
